requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.1
django-allauth==0.51.0
django-bootstrap-form==3.4
django-bootstrap-v5==1.0.11
django-filter==22.1
django-friendship==1.9.6
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
twilio==7.15.1
types-cryptography==3.3.23
tzdata==2022.5
urllib3==1.26.12
whitenoise==6.2.0
